magic development - magic schools system
added a magic system which divides the types of spells the player can cast into six schools.  also added a mechanic where when the player adds a tower, it creates an event that makes you assign a wizard type to the tile.  wizard type in tile now increases matching magic school int